+++ console/src/new-components/Switch/index.tsx
@@ -1,0 +1,1 @@
+export { Switch } from './Switch';